Remove VW2CremeRegressor from compat
